Reimplementation of IMGputBitFont
I've converted the font routine to C++. I believe there might be a
signed/unsigned issue - I made AX and bitMask both unsigned chars,
when maybe I should have just let them be ints. The reason I'm
concerned is that the prototype defines the bitmap buffer as a char*,
which the compiler should interpret as a signed char.