Fix comparison between different enum symbols. Starting point: OTERRAIN.cpp: In static member function 'static short int TerrainRes::max_height(TerrainTypeCode, SubTerrainMask)': OTERRAIN.cpp:176: warning: comparison between 'enum TerrainTypeCode' and 'enum<anonymous>'
TOTAL_TERRAIN_TYPE is an anonnymous enum just defined as the total number of
TerrayTypeCode's.

It may be fixed by including an special entry in the TerraintTypeCode enum
to signal the last value (or the number of values, if all of them are
sequential). The "easy" way to avoid the warning is to cast to the tagged enum type.