Audio: Correct volume adjustment functions
When wavs are spawned, they are attenuated by the per-wav volume and the global
Audio::wav_volume.  When the volume of an existig wav is changed by the per-wav
volume_*_wav(), it is *not* attenuated by wav_volume.  When the global
set_wav_volume() is used, it assumes that all wavs are already attenuated by
wav_volume and adjust their volume accordingly.

This behaviour is used by the gui volume control code, which first calls
set_wav_volume() on all wavs and then volume_*_wav() on the music channel to
make music volume independent from everything else.  This isn't very good
behaviour and should probably be fixed later.

Also, the logarithmic-linear conversion now gives 0dB == 1, -6dB == 0.5.