War factory bug patch
Hello, I want to send a patch for fixing a bug relating war factories,
where production can be cancelled by the player if the building has a own
spy or a phoenix is near. This shows the whole info for the war factory
(which is ok), but also shows the cancel button, which is a bug. This is
discussed in the forum on this thread:
http://www.7kfans.com/forums/viewtopic.php?f=13&t=640