GameMP: Abort setup if mp_join_session() fails.
mp_join_session() is ultimately responsible for setting mp_obj.host to a
valid value. If it fails, enet_host_service() will be called with a null
host, resulting in a crash. This can happen if, for example, the player
cancels the password input box when joining a game. Returning to the
main main if mp_join_session() returns 0 will ensure that we don't
continue the setup with an invalid state.