ONATIONB: Fixed a bug with Alliance visibility and new nations spawning
When a new nation spawns, it sets its relations with others and others
with it. However, the NationBase member variable is_allied_with_player was
also reset for each already existing nation in init_relation.
is_allied_with_player is used for visibility-checking, and so a new nation
spawning caused all existing alliance-visibility to vanish.