Town Network: initial files and algorithm implementation.
Includes the basic definitions of a Town Network and the TownNetworkArray,
which is to TownNetwork what TownArray is to Town. Extended Town to
include properties pertaining to Town Networks and handle events relevant
to Town Networks.

Updated by Jesse Allen <the3dfxdude@gmail.com> to merge cleanly, and to
update the automake files.