Unit: stop building firm when nation is out of cash.
Before this change, if a nation was out of cash then the units would wait on the spot for money to be available, after which construction would commence.
The effect on AI actions is not entirely clear, except for that it is an improvement: actions and units are freed up faster.