MultiPlayer: Let enet manage the list of players.
player_pool becomes a list of authenticated players for the game session.

poll_players is implemented again as a way to refresh player_pool.

pending_pool was added as a way to track pending connections.

playerIds are generated and no longer refer to index values of any particular list.

The displayed player list is refreshed more effectively based on connection events.