gettext: Create wrapper to gettext convenience header.
Two problems have been seen with the currently implemented convenience header.
1) Does not compile on mingw32 with #undef ENABLE_NLS.
2) The header in the file remarks a GPL3+ license, but upstream actually has an LGPL 2.1 license.

Due to the unattributable license change, the source here cannot be verified. We need to make a clean break back to the original GNU gettext.h, if we want to use it.

In this patch gettext.h was renamed to gettext-gnu-old.h. A new wrapper was added to source the implementation of gettext-gnu-old.h. Subsequent patches will add the upstream gettext.h as gettext-gnu.h. Any project customizations will be added to the gettext wrapper. Finally, the old header will be deleted after all appropriate fixes are in place.