ErrorControl: Fix non-portable cast.
This should enable 64-bit-to-64-bit multiplayer matches. I also reviewed ErrorControl, and I didn't see any other casts as found here. There are usages of long in the encoded messages, which will affect 64-bit-to-32-bit, but that is a problem for another day.

ErrorControl is a bit redundant in a couple aspects, and could simply be removed, with some testing. It was likely used in situations that are no longer relevant.