MultiPlayer: Remove extra padding in new MpMsgs.
Originally the structs were being aligned with their end to a multiple of 8 bytes despite what the compiler done to ensure extra strictness at the ending address. This isn't needed as the stride is defined around the size of the largest basic storage type, which all are 4 bytes. We can use -Wpadded to check for the actual requirement.

Further note: All MpMsg structs are assumed packed with no adjustment necessary by the compiler. All padding that could be done is marked so it is tracked. If the struct is not packed correct on a compiler for some reason, please let me know. Also note the data types are non-variable size.