MultiPlayer: Add 7kfans.com as a multiplayer service.
This is far enough along to begin testing.

The login procedure is simple. Open your browser to 7kfans.com/forums, and log in if you are not logged in. If you keep your browser open, you may have to refresh the page.

Start the game, and go to the multiplayer menu. Select 7kfans.com from the menu and enter your forum username. Then you will be able to host or join games mediated through 7kfans.com.

This may be continue to be a work in progress, but it passes my initial implementation effort.