Unit: Disable fake spy loyalty.
This change intends to help the AI since they do use a strategy to announce betrayal. But because the constant loyalty only happens with a spy, it always reveals to the cloaked nation who is a spy. Human players can easily weed out spies, making the AI more ineffective. This should improve AI difficulty some.

This also can encourage the announce tactic in multiplayer matches. Previously, no one with knowledge of this would ever consider this since it is an definitive death of your spy when against another human player.

This also happens to align the spy behavior with 7k2.