Fix bug that creates extra music selection buttons
Mentioned here: https://www.7kfans.com/forums/viewtopic.php?f=13&t=813.

This is because MAX_RACE was originally 7 but updated to 10 with the
release of the AA expansion. This value was also used in the loops
generating the buttons to select music tracks. New music was not
added to the game.