Nation: Implement better penalty levels for civilians killed.
The motivation for this change started when someone showed that a nation can be caused to have a large rebellion when sending workers of the same race to attack. While this directly addresses the Nation::civilian killed function, this also changes how civilians are viewed in Unit::hit_target and Unit::is_civilian. If the change were to simply disallow sending civilian-like units, as the function previously penalized, then the code for hit_target on the civilian penalty is dead code. I highly suspect that the code originally intended a penalty, but overtime, it got restricted. What needs to happen is to have improved levels for the penalty. It is easy to detect town defenders using Unit::in_defend_town_mode, rather that the hodge-podge skill<20 or other factors that don't seem to work.

In testing the code as it was, I realized that in high resistance towns, most of the town's residents will pop out to defend the town. This mean that a more aggressive player simply would be better off attacking the town out-right, if he wanted to destroy it, since he would not get penalized as much as he would if he waited to influence the town by other means. The low resistance town, may mean more in-town deaths. So I felt that the effect of attacking a town with high resistance should at slightly higher penalty for the number of deaths, by accounting for the town defenders dying. But since -3 is very large in general, it can't be that high, as I want to keep a balance somewhat to what players are used to. Therefore, I took a bit of a hint from 7k2, and I set in-town deaths to -2 loyalty of that race, and -1 if a defender dies.

Lastly, I thought about recruited units, mobilized units, and re-appropriated town defenders. These units (if non-combat) caused a big penalty as was noticed before for the "attacker" nation, and a somewhat of a penalty for the "defender". I though it would be appropriate to remove the loyalty penalty all-together, and keep a split -0.3/-0.3 reputation penalty for both sides to share for having civilians involved in battle.

One thing to note, is the loyalty penalty is applied to all firms, towns, and mobile units immediately, which is why it has such a big impact. Reputation is not immediate, but affects loyalty over time, since it factors into the target loyalty equation. So aggressive players take note. Attacking towns will bite a bit more than in the past, while not enough to cause rebellion, but you'll have to keep your people of the dying race happy more. Then you also have to heal your reputation. Granted, you can go quite negative before you need to worry.

Players ought to report results. There should be a slight loyalty and larger reputation impact for being overly aggressive, and a slight reputation impact for loosing civilians. It is pretty much impossible to cause an instance rebellion, other than you attacking several towns or killing alot of civilians purposefully. This could slow down aggressive players slightly, but not too bad, if it does.

The following is the table showing the change:

(Loyalty Loss,Reputation Loss)
Attacker (or Nation causing the death)
In-town: (-3,-1) -> (-2,-1)
Town-Defender: (0,0) -> (-1,-1)
Civilian: (-3,-1) -> (0,-0.3)
Caravan: (-3,-10) -> (-3,-10)

Defender (or Nation inflicted with death)
In-town: (-1,-0.3) -> (-1,-0.3)
Town-Defender: (0,0) -> (0,-0.3)
Civilian: (-1,-0.3) -> (0,-0.3)
Caravan: (-1,-3) -> (-1,-3)