Help: Fix signed vs unsigned comparison in FPU.
There is an issue where the type used for the mouse_still_time caused an incorrect comparison, and would instantly trigger help popups without there being any wait. The issue is that the comparison in Help::should_disp used a float number, which cause all the other numbers in the expression to upcast to float. The return from Misc::get_time is unsigned long -> float, and then mouse_still_time signed int -> float.

When Misc::get_time on Windows is executed, it uses a time value that begins at system boot. If your machine is on for a very long time, like mine are, there is potential for this time value to get large (negative when cast), and eventually roll over. When the expression is compiled in gcc, the instruction to load an unsigned long for current time is used, and an instruction to load the signed int is used, which will always create a negative vs positive comparison in the FPU.