Autotools: Use pkgconfig inputs if possible and make portable windows builds easier on mingw-w64.
With now that we have C++11 compliance in mingw-w64, the default gcc build automatically links in a pthreads compliant library. Even when the application doesn't use threads, it gets linked in by default. If you use libraries provided by MSYS2 repositories, they are also linked with the library. It was necessary to provide externally provided libs that were built differently, as well as using a potential hack to the linker to get the final executable to drop the dependency. It is harder to remove the dependency when the application is not threaded.

See also this:
https://stackoverflow.com/questions/13768515/how-to-do-static-linking-of-libwinpthread-1-dll-in-mingw