TOWNBTYP.dbf: Fix stored size of the HOUSE town type.
The problem is that the size of HOUSE type was being stored, and incremented by one erroneously. This was essentially impossible to trigger a crash in practice, but when the game auto generated the layout of the towns, it iterated through a list size of 121, and considered an invalid record '121' past the end of the town_build_array. Due to what always followed in memory, the town build 121 was always rejected. If it hadn't got rejected, bounds checking elsewhere in the code would have caught it if it was used, since the game does keep track of the actual number of records in a database.

This problem has been present in all versions of 7K, 7KAA, and 7K2.

Reported by t0suj4.