RaceRes: Don't store translated race name and adjectives.
NOTE: The convention for Maya/Mayan will be:

Maya: Both adjective and proper name for referring to a person, people, and nationality
.
i.e. Maya King, Maya Nation

Mayan: OK as an adjective referring to anything that isn't a person,
 people, and nationality.
i.e. Mayan scroll of power, Mayan culture, Mayan language

The goal for this is translatability, not word-pickiness. This convention matches the original game design pretty well, with just one change coming on the unit profile.