Reintroduce 'none' option for video & input, united under one makefile/configure switch 'ui_backend'.
Note that the C++ source allows switching input and video between none/sdl independently.
The automake build system, however, only has the option to switch both at the same time.
We could've allowed independent switching in the makefile as well, but that didn't seem as useful -- building 7kaa as 'none' is more for verification that the switching still works;
the end-game is to allow new 'tester'-like projects to compile with none for one or both, which projects will most likely have their own configuration.