VgaNone: Add an actual surface buffer to make it usable when running parts of the game code.
The game currently crashes when running code with VgaNone, because it doesn't have a buffer.
This adds a 'dummy' surface buffer to which the game can read and write.
The contents of the buffer will currently be useless (all 0's) because of the dummy palette.
It could be extended with some kind of palette such that you could glance at the contents of the back and front-buffer.