ReplayFile: Serialize more necessary data for to enable full replay capability.
Looks like we'll need this to debug desyncs. First pass was a packet dump. Now we provide all game starting information and preserve entire receive queue frames.

This will likely be considered experimental for a long time, but it's really quite neat to see this in action now.

Note: A limitation is that we won't be able to record replays starting from a save file. We would need to serialize a whole save file to do that.