Edited some messages for better clarity.
Changed "Bribing Succeeded/Failed to Bribe... This changes the form of the word to match "assassination" as it is not "assassinating."

Changed the spacing on a word and number combination. There is enough space to display the text with spacing, and "properly" the text would include it.

Changed a comma to a period. instead of having "proceed" hanging off the previous sentence, i added the period in question because if the proceed clause wasnt elliptical, it would be a full sentence, thus requiring the period. the previous usage is more commonly used for modifiers or opposition instead of a full question.

Changed form of word "fail" to "failed". It was morphological incorrect.