Caravan: Remove global variables for the stop goods selection interface.
Because the state of the stop goods selection were in global variables, the game code went to great lengths to synchronize the buttons with the selected caravan. This included blasting a message in multiplayer to each player of each game client's current trade caravan selection. This had consequences of requiring an additional frame for synchronization, and other issues like the buttons popping back out momentarily, or even dropping the action altogether if it wasn't in sync with the other clients. In a rare case, it could even crash the game.

7k2 dropped the global variables, and showed that the goods info could be pulled directly from the TradeStops at display time. There is no reason to synchronize an extra copy in the main loop. This change mimics the improvement of 7k2, but keeps the same selection behavior and interface of the original game.