systematic period usage
if sentence is complete, a period is used. Otherwise, periods are not used. Note an incomplete sentence will use a period if there are multiple sentences in the string. i was quite confused about how to use periods. this is done to hopefully allay confusion in regards to using periods. 

for example, in OSaveGameProvider, there was previously:

Path too long to the saved game.
Path too long to the saved game