Don't always call box.msg on a GameFile error.
Calling box in auto save and during multiplayer is a bad idea, since it is a blocking operation. There is already news feedback on successful save in multiplayer. The only place it might make sense to give messages are in the SaveGameProvider. So make the GameFile status reporting contingent on whether it is an appropriate time to see it.