Standardize on 387 floating point unit to support stable multiplayer.
Depending on arch, gcc uses sse on new targets and 387 on old targets. This is generally on 32-bit vs 64-bit x86. They are not compatible in math operations though. We need require the same as much as possible to keep sync.

Since non-x86 is not guaranteed, add appropriate detection and warnings to the user.