GameMP: Process MPMSG_END_SETTING anyway when client hasn't seen the MPMSG_START_GAME from the host.
The game tries to sync the clients on receiving the final settings before starting the game. When the host broadcasts the start message, some clients may process the message before others see the start message. This causes some clients to assume there are issues when there is not any.

This patch uses the 7k2 approach of processing the message, even though it's not foolproof still. An alternative would be to just not attempt to synchronize and delete the code, since the messages are guaranteed anyway, and let downstream or the packet layer deal with the connection. But for now we will try the 7k2 approach.