Unit: Don't dereference a Firm that is not a Camp.
Before previous commit, soldiers may still be attached to their Camp that was deinit. If the slot for the Camp was replaced by another building, it could corrupt memory.

While it is unlikely to happen when the defense flag is cleared, this puts a proper check that the fort is really still there.