FirmCamp: Don't close camp if all soldiers are mobile.
The AI considers forts unattached to towns not useful, so they are flagged for closure. During the periodic processing, if it was flagged to close, and there are no soldiers in the building, it will be sold then. This is a problem because the soldiers might be doing something useful, and now they have no fort.

There are reasons why the AI might have constructed a fort that is unattached. The basic reason is due to a geographic obstacle. There is no room to place a fort to one side of the town and the separation is too large. A bigger reason is the AI actually tries to settle separate regions, if they are large enough and sees an opportunity. Regions in the game can be completely isolated continents.

In a case observed, the AI settled an island by first bringing soldiers, and building a fort to hold them, of course unattached to a town, since people have not been brought. There was a fryhtan lair, and due to high probability of an attack from that lair, the lair attacked to fort. The soldiers went on defense, and then the AI immediately sold the camp. This wreaked havoc on the initial goal of settling the island, because then another needs to be built, and then another attack may come, and the issue repeats. The AI also tries to send peasants to settle a town next to the fort to fully establish a presence, but this too is hampered.

The fix is to check that the camp is empty of soldier both within and mobile. This does not include the general, because if it was only populated by a general, and all soldiers die, this would be an opportunity to relocate to a more useful location.

The added benefit, however, is even if the fort is in a populated region, just unattached, it doesn't get immediately closed if the soldiers are in use. This may prevent other situations where the AI has alot of mobile soldiers, and seemingly not enough forts.

The fix when tested on the AI island migration case the town settling process improved dramatically, even when attacked during establishment. This may improve the competency of the AI on sea missions, as that seemed to be lacking in the classic version.

Note: 7k2 will sell the camp under the same basic check of having no soldiers in camp, which means the issue of closing camps too quickly could exist there. The should_close_flag, however, is enabled only as long there is still room for all soldiers commanded. If there is a soldier capacity handling issue in 7k, we will address this in a more appropriate test case, but it could be we addressed the original issue well enough.