Unit: Prevent invisible unit from trying to walk onto map.
In a test game, an AI player was doing poorly. A troop team was headed to a ship when the king died. The general of the troop became king, but some of the troop made it to the ship. The new king got surrounded and fled. The AI sent the the king to a camp, along with his immediately command troop. The order to go to camp was processed for a soldier on a ship. This caused the soldier to path find from x=-1, attempt to occupy a slot on x=0. When the ship finally landed in the completely wrong spot (an island), the soldier's loyalty was zero. Seeing no one else to join, the soldier immediately resigned. Sometime later, a unit tried to path find through the x=0 slot the soldier erroneously occupied. The game crashed because the soldier was not found.

Added checks for assigning and attempting move an invisible unit. If this happens, assign returns without assigning.

The proper thing to do would be to detect the ship status, and eventually process the ship back to the place where the soldier is assigned. This requires more work.

The AI issues will persist with ordering units that are on ships.