Firm: Add request builder ability.
A button will show in place of the hammer for a construction worker, if there is no construction worker and one sufficiently idle one is available to send. Clicking on it will order the closest construction worker to the building for repair.

The current button used is considered temporary until an image is added that is more appealing.