Town/Firm: Correct checking for room to add structure links.
The original game devs set limits the number of links per type of link and knew that if the limit was exceeded that the link could not be made. However in normal game play, the exception will not get thrown and caught, corrupting the game state for when the structure is deleted. They tried to prevent this from occurring but the check did not prevent partial linking and they also introduced a bug corrupting town linking to firms at a lower limit.

This patch corrects the check and allows the game to continue without corrupting the links. If there is no more room for a link on a structure, then it does not start the linking process in that case.

The main bug this was discovered on, incorrect check on max number of town links on a firm, and the problem with partial linking also exists in 7k2.