Display credits for the project
Many of the prolific contributors added. If you are a contributor, please state what you want printed on the credits page, otherwise it will based on commit history. It is alphabetical to last name.

The bible script font, while used in the game, we don't have the font resource file provided yet. Use the news font for printed text for now.