Town: Fix reporting recruit loyalty decrement.
The AI uses the possible loyalty decrement to know if it should recruit from a town or not. Unfortunately it is never really correct if they are recruiting many in a loop.

The loyalty decrement is supposed to range from 1-10 in integer values, but the accumulated recruit penalty is not accounted when reported to the caller. This means the function was only ever reporting 1-5, which is rarely correct. Also, only integer truncation is performed, but it should really round up to be safe. Recruit heavy personalities are more likely to trigger rebellions.

The fix returns the fully calculated loyalty decrement that will happen, in integer form rounded up. This should mean that the ai will never accidentially trigger rebellion by using the recruit function. Play testing shows that the accidental rebellions stopped. Testing over 100 matches suggests ai-only matches increased in playtime by 1 year std deviation.

7k2 has the same problem in reporting the decrement value and then using it to decide to recruit.