Unit: AI team help request should not apply to units in fort
If request goes through, then path finding is called, and bad things will happen. It seems to only be triggered on generals in the cases I found. Probably because it always gets attached to its team for defending purposes.

The unit should be mobilized first (if this is even a good idea), but this ability does not exist here. For now, do not apply the request.