Town: Fix cases of unjob_town_people() leaving workers in limbo.
The internal function unjob_town_people() was leaving workers in limbo in two cases:
1) When picking a unit to kill, it would unjob a worker, but erroneously used the Firm::resign_worker return code, and end up not killing the worker, now "fired".
2) When picking rebels, it could unjob a spy worker, but then leave the spy in town as they can't rebel, leaving the spy worker "fired".

These things need to be fixed, as they can be considered exploitative. But this could be somewhat of a devastating fix, as one of the most common scenarios, earthquakes, or the town getting attacked, means this will cause much more damage more quickly to towns that have workers than before.

A similar function in 7k2 was rewritten and does not use the resign_worker return code. I'm not sure how worker spies were accounted for.