Unit: Finish attack move if firm is deleted.
Every attack order is really two orders. The first order is to move to the target if out of range and the second is to attack the target. The game code in the process attack functions is possibly to strict about objects being deleted to clear the order. This immediately removes any knowledge of what the target was supposed to be, and stops the unit. For the AI, the idle code kicks in, and if nothing is around, issues an immediate return order. For the player, in the rarer case they issued the attack order from a long distance, you'll probably have your units end up scattered across the map that now you have to clean up.

This change is primarily to help the AI be more effective. For one, when the firm is deleted, the move is finished and idle code kicks in which does a simple, but reasonable job at finding more targets. For the player, the units will at least go where you want them to go.

I suspect most RTS is designed around this kind of desirable behavior.

7k2 attack processing is quite different here, and perhaps a good cleanup of unit action implementation, and hopefully allows proper attack-move.

This is part of a series of patches to improve attack-move.