Town: Implement town peasant migration.
Peasant migration was intended similar to worker migration but because of a coding error, it was never triggers. The code was clearly not fully implemented due to lack of testing.

The original code, if enabled, fires too quickly, and allowed things that said was not supposed to happen in the code comments. The rate think_migrate() runs is now reduced to 1 in 15 days per town. Finally the rules have been updated to rate the town migration choice like workers do. The chance of migration seems almost as low as in worker buildings. More testing will be needed to see if we need to lower the chance at execution.

The original code also allowed migrating over long distances and also can let peasants get lost along the way. This seems ill advised, and provides no meaningful benefit to the game play, loosing peasants over time or favoring one reputable player of the desired nation. It also destroys the independent town "mix" of the original game design. This code is disabled.

The peasant migration option is interesting as it seems like it could help balance out nationalities in your own or ai towns for positive impact. There may be also new ways to play the game knowing how to leverage this.