FirmMarket: Report and give control on restocking sources.
The game already had given the ai control which type of building to source material from. The flag would only work for the ai, and it wasn't consistent on how it was used. The ai never exactly needed it either, it appears the game authors were originally trying to use the firm links instead. Indeed, none of the scenarios actually had any "retail" flag set that I could find. I believe that the variable was a left over from early prototyping, then employed in a patch after the release of ancient adversaries to fix some quirk on preferring mines over factories. (comment indicates there was a patch to use it)

This change exposes the flag in the market info, and gives a button to switch the type for the user. There are four choices, and it does seem depending on the choice you can actually gradually clear out one type without resorting to wiping it out completely. Lastly, the code was changed to make sure filtering is applied even if not ai. This means that whatever the ai uses will be preserved when you take them over. The ai should also now update the type on it's own, and not just when it builds the market.

The enhancement was suggested by MicroVirus and sraboy.