LocaleRes: Set the LC_MESSAGE environment variable when the system does not support the setlocale setting.
Gettext assumes usage through environment variables when the system is not POSIX compliant. This also defaults the language to English on a related issue, because when we can't figure out the default system language, we can't then select the proper font.

The user has more options to set their language locale. The ConfigAdv setting for locale will now work on non-POSIX settings. The environment variable SKMESSAGES is now available non-POSIX situations as well.

This patch also does more prepping for a future menu to select the language in game.