Spy: Try to save friendly cloaked spies in firms and towns when an enemy nation is defeated.
Calling Spy::change_cloaked_nation() calls Unit::spy_change_nation() which is only coded for mobile units. This is probably not intended. At least now try to save the units by mobilizing, and then switching cloak, just as the original comment intends. In the case the unit can capture the firm, it will do it.

There was an assertion that will access a garbage pointer if the Spy::drop_spy_identity is used. This is because that function calls SpyArray::del_spy. This is corrected.

It is likely this code was never properly tested and possibly trashed things along with it. But because it could happen in a live game when Nation::defeated() is called (usually AI), we'll improve it. There is still more that needs looked at.