Unit: Remove action_para assertion.
action_para is allowed to be zero at least entering process_idle, as we
need to ensure that units that have died have the reference cleared
before some other function tries to intepret or reuse the unit record
number.  So in the first case, it has to go. The other checks may or may
not be useful, but that is a web I cannot untangle right now.