FirmMarket: Fix distance check.
Spotted this bug while stress testing code. The y axis is completely
zeroed out, which would have the effect of allowing markets unrestricted
to one half of the map, if the AI otherwise wants to. This at least keeps
the distance a little closer to home. Maybe the AI will be a little more
reasonable on building export markets.

7k2 has the same bug.