NationAI: Prevent recruiting when there is no overseer.
Due to a logic bug, the AI could recruit civilians when the action was
to move to another town when there was no overseer and without the
minimum recruit loyalty. This would trigger rebellions. The logic is now
fixed.

7k2 has the same fix. Although using the town 'has_linked_own_camp' flag
is unreliable. The flag is updated in the town periodic processing, which
may not happen before the recruit logic is processed. However this issue
is found all over the place and not as severe an issue. It maybe better
to replace these with Town::has_linked_camp(..., needOverseer=1) in the
future.