Misc: Add rects_distance() function.
This function can be used over the points_distance() function when
dealing with rectangular shaped objects in an integer coordinate system.
This will handle finding a center coordinate that makes sense when a
precise distance from the center of an odd sized shaped to an even sized
shape is not possible.