Switch distance calculation used for firm and town linking.
When placing alternating firms and towns, the points_distance() function
is called with the center x,y member variables to link the towns. This
produces an effect that depending on the side of the town, the max
distance between firms and towns changes. If you want towns to link for
management purposes, you have to squeeze them together closer, meaning a
very cramped layout or only building in a single line of towns and firms.

The change here makes the distance consistent on all sides of the town,
linking as expected during game play.