Update AI and generation code distance measurements.
This updates AI decision making using precise measurements where needed
and also map generation. Not all AI distance measurements are updated,
as the effect of being off (like looking for a raw site) may be
negligable in practice. Anytime it is critical to linking, bldg layout,
or in battle, this is what we are changing here.

The function TownArray::find_nearest_town is not used, and it seems to
strongly imply that it is from the Capitalism game. Towns were likely
measured to their "center" being a center of town, but there were not
generated with a regular shape, so it was arbitrary. It is updated in
case it does get used again in 7k.