TownAI: Pick a better direction to scout.
The original choice requires a min 50 distance in both x and y strongly
biases the move order to be exactly to a map corner, especially since
it could go up to 50 more in x and y again.

The random choice now is simply picked with-in one edge of the map
randomly which is what most people do anyway. Hopefully that improves
the change enough the AI can connect with other players very early in
the game.