Unit: Stop ordering blocking idle teammate to action destination.
This order to an idle teammate to push to the same destination only
serves to obstruct completing the action. I think there was team
movement in mind here, but if the units are not ordered to move
together, then an individual move order shouldn't be swapped to
someone else not going there.

The fix is to use the more usual block-by-idle unit resolution.